Refactor backend.py into multiple files to support shlo and torch compiler
